New dynamo

+++ tastypie_dynamodb/resources.py
@@ -1,3 +1,4 @@
+import copy
 from django.conf.urls import url
 from django.http import Http404
 
@@ -5,9 +6,7 @@ from tastypie.exceptions import NotFound
 from django.core.exceptions import MultipleObjectsReturned
 from tastypie import http
 from tastypie.utils import dict_strip_unicode_keys
-import boto.dynamodb
-from boto.dynamodb.condition import EQ, ConditionTwoArgs
-from boto.dynamodb.exceptions import DynamoDBKeyNotFoundError
+import boto.dynamodb2
 
 from tastypie.resources import DeclarativeMetaclass, Resource
 from tastypie_dynamodb.objects import DynamoObject
@@ -50,20 +49,49 @@ class DynamoHashResource(Resource):
 
     def __init__(self, *a, **k):
         super(DynamoHashResource, self).__init__(*a, **k)
-        self._hash_key_type = int if self._meta.table.schema.hash_key_type == 'N' else str
 
+        # There is a bug in boto, which doesn't assign proper
+        # data_type to table.schema fields. We correct this here
+        schema = self._meta.table.describe()
+        field_defs = copy.deepcopy(schema['Table']['AttributeDefinitions'])
+        self.table_schema = copy.deepcopy(self._meta.table.schema)
+
+        for orig_field in self.table_schema:
+            fdef = filter(lambda field: field['AttributeName'] == orig_field.name, field_defs)[0]
+            orig_field.data_type = fdef['AttributeType']
+
+        # Get data_Type of hash key
+        self._hash_key_type = int if self._get_hash().data_type == 'N' else str
+
+        # Get list of all indexed fields
+        self._meta.indexes = {}
+        for index in self._meta.table.indexes:
+            # Get all indexes and then find RANGE in there
+            self._meta.indexes[index.name] = filter(lambda part: part.attr_type=='RANGE', index.parts)[0].name
+
+    def _get_hash(self):
+        tmp = filter(lambda field: field.attr_type == 'HASH', self.table_schema)
+        if tmp:
+            return tmp[0]
+        raise Exception('Couldn\'t find HashKey!')
+
+    def _get_range(self):
+        tmp = filter(lambda field: field.attr_type == 'RANGE', self.table_schema)
+        return tmp[0] if tmp else None
 
     def dispatch_detail(self, request, **k):
         """Ensure that the hash_key is received in the correct type"""
         k['hash_key'] = self._hash_key_type(k['hash_key'])
 
+        if 'range_key' in k and type(k['range_key']) is str and k['range_key'][-1] == '*':
+            return self.get_list(request, **k)
         return super(DynamoHashResource, self).dispatch_detail(request, **k)
 
     def resource_uri_kwargs(self, bundle):
         kwargs = { 'api_name': self._meta.api_name,
                  'resource_name': self._meta.resource_name }
         if bundle:
-            kwargs['hash_key'] = getattr(bundle.obj, self._meta.table.schema.hash_key_name)
+            kwargs['hash_key'] = getattr(bundle.obj, self._get_hash().name)
             if kwargs['hash_key']:
                 kwargs['hash_key'] = str(kwargs['hash_key'])
 
@@ -74,29 +102,48 @@ class DynamoHashResource(Resource):
             url(r'^(?P<resource_name>%s)/(?P<hash_key>.+)/$' % self._meta.resource_name, self.wrap_view('dispatch_detail'), name='api_dispatch_detail'),
         ]
 
+    def get_dynamo_filter(self, kwargs):
+        filt = dict()
+        filt[self._get_hash().name] = kwargs['hash_key']
+        if self._get_range():
+            filt[self._get_range().name] = kwargs['range_key']
+        return filt
+
     def _dynamo_update_or_insert(self, bundle, primary_keys=None, force_put=False):
-        primary_keys = primary_keys or {}
-
         bundle = self.full_hydrate(bundle)
-        item = self._meta.table.new_item(**primary_keys)
 
-        #extract our attributes from the bundle
+        if primary_keys:
+            filt = self.get_dynamo_filter(primary_keys)
+            # Extract primary keys
+            if force_put:
+                item = filt
+            else:
+                item = self._meta.table.get_item(**filt)
+                if not item.values():
+                    raise Http404()
+        else:
+            # An attempt to create a new item
+            item = dict()
+
+        # extract our attributes from the bundle
         attrs = bundle.obj.to_dict()
 
-        #loop and add the valid values
+        # loop and add the valid values from the given bundle
+        # to the dynamo item
         for key, val in attrs.items():
             if val is None:
                 continue
-
             item[key] = val
 
-        #if there are pks, this is an update, else it's new
+        # if there are keys, this is an update, else it's new
         if not primary_keys or force_put:
-            item.put()
+            # New or PUTting item
+            self._meta.table.put_item(item, overwrite=force_put)
         else:
-            item.save()
+            # Save and overwrite if item exists already
+            item.save(overwrite=True)
 
-        #wrap the item and store it for return
+        # wrap the item and store it for return
         bundle.obj = DynamoObject(item)
 
         return bundle
@@ -111,18 +158,18 @@ class DynamoHashResource(Resource):
 
     def obj_get(self, bundle, request=None, **k):
         """Gets an object in Dynamo"""
-        try:
-            item = self._meta.table.get_item(consistent_read=self._meta.consistent_read, **k)
-        except DynamoDBKeyNotFoundError:
+        filt = self.get_dynamo_filter(k)
+        item = self._meta.table.get_item(consistent=self._meta.consistent_read, **filt)
+        if not item.values():
             raise Http404
-
         return DynamoObject(item)
 
     def obj_delete(self, bundle, **k):
         """Deletes an object in Dynamo"""
-
-        item = self._meta.table.new_item(**k)
-        item.delete()
+        filt = self.get_dynamo_filter(k)
+        item = self._meta.table.get_item(consistent=self._meta.consistent_read, **filt)
+        if item.values():
+            item.delete()
 
     def patch_detail(self, request, **kwargs):
         deserialized = self.deserialize(request, request.body, format=request.META.get('CONTENT_TYPE', 'application/json'))
@@ -153,40 +200,39 @@ class DynamoHashResource(Resource):
         pass
 
     def get_count(self, attr_filter={}):
-        if self._meta.table.schema.range_key_name:
-            attrs = [self._meta.table.schema.range_key_name,
-                     self._meta.table.schema.hash_key_name]
-        else:
-            attrs = [self._meta.table.schema.hash_key_name]
+        # if self._get_range().name):
+        #     attrs = [self._get_range().name),
+        #              self._get_hash().name)]
+        # else:
+        #     attrs = [self._get_hash().name)]
 
         dynamo_filter = {}
         for key, val in attr_filter.iteritems():
-            dynamo_filter[key] = boto.dynamodb.condition.EQ(val)
-        _items = self._meta.table.scan(scan_filter=dynamo_filter,
-                                       count=True)
+            dynamo_filter[key + '__eq'] = val
+        _items = self._meta.table.scan(**dynamo_filter)
 
     def get_uri_list(self, request, attr_filter={}):
         """ Gets a list of resource URIs of all objects in this table"""
-        if self._meta.table.schema.range_key_name:
-            attrs = [self._meta.table.schema.range_key_name,
-                     self._meta.table.schema.hash_key_name]
-        else:
-            attrs = [self._meta.table.schema.hash_key_name]
+        # if self._get_range().name):
+        #    attrs = [self._get_range().name,
+        #             self._get_hash().name]
+        # else:
+        #    attrs = [self._get_hash().name]
 
         dynamo_filter = {}
         for key, val in attr_filter.iteritems():
-            dynamo_filter[key] = boto.dynamodb.condition.EQ(val)
+            dynamo_filter[key + '__eq'] = val
 
         # TODO do a query if filter HASH_KEY available
-        _items = self._meta.table.scan(scan_filter=dynamo_filter,
-                                       attributes_to_get=attrs)
+        # TODO see if you can limit attributes that you get
+        _items = self._meta.table.scan(**dynamo_filter)
 
         def hash_uri(item):
-            return '%s%s/' % (self.get_resource_uri(), item[self._meta.table.schema.hash_key_name])
+            return '%s%s/' % (self.get_resource_uri(), item[self._get_hash().name])
         def range_uri(item):
-            return '%s%s/%s/' % (self.get_resource_uri(), item[self._meta.table.schema.hash_key_name], item[self._meta.table.schema.range_key_name])
+            return '%s%s/%s/' % (self.get_resource_uri(), item[self._get_hash().name], item[self._get_range().name])
 
-        if self._meta.table.schema.range_key_name:
+        if self._get_range():
             items = [range_uri(it) for it in _items]
         else:
             items = [hash_uri(it) for it in _items]
@@ -203,55 +249,69 @@ class DynamoHashResource(Resource):
 
         # Try to filter by hash_key, if provided
         # Or filter by kwargs['hash_key'] - this happens when we get a wildcard range key URI request
-        hkey = self._meta.table.schema.hash_key_name
+        hkey = self._get_hash().name
+
         if hkey in request.GET or 'hash_key' in kwargs:
             hkey_in_next = True
             value = request.GET.get(hkey, kwargs.get('hash_key', None))
-            dynamo_filter[hkey] = boto.dynamodb.condition.EQ(value)
-            hash_key_value = request.GET.get(hkey, kwargs.get('hash_key', None))
+            hash_key_value = value
+            dynamo_filter[hkey + '__eq'] = value
 
         # Exclusive start key - when offset is required
-        esk = []
+        esk = {}
         if 'offset_hash' in request.GET:
-            esk.append(request.GET['offset_hash'])
-            if self._meta.table.schema.hash_key_type == 'N':
-                esk[0] = int(esk[0])
+            hash_offset = request.GET['offset_hash']
+            if self._get_hash().data_type == 'N':
+                hash_offset = int(hash_offset)
+            esk[self._get_hash().name] = request.GET['offset_hash']
 
-        if self._meta.table.schema.range_key_name:
+        if self._get_range():
             # We are dealing with a range table!
             if 'offset_hash' in request.GET and 'offset_range' in request.GET:
-                offset_hash = request.GET['offset_range']
-                if self._meta.table.schema.range_key_type == 'N':
-                    offset_hash = int(offset_hash)
-                esk.append(offset_hash)
+                range_offset = request.GET['offset_range']
+                if self._get_range().data_type == 'N':
+                    range_offset = int(range_offset)
+                esk[self._get_range().name] = range_offset
 
             # a 'range' table, let's try filtering
-            rkey = self._meta.table.schema.range_key_name
+            rkey = self._get_range().name
             if rkey in request.GET or 'range_key' in kwargs:
                 value = request.GET.get(rkey, kwargs['range_key'])
                 if value != '*':
                     if value[-1] == '*':
                         # wildcard filer, we need begins_with
-                        dynamo_filter[rkey] = boto.dynamodb.condition.BEGINS_WITH(value[:-1])
+                        dynamo_filter[rkey + '__beginswith'] = value[:-1]
                     else:
-                        dynamo_filter[rkey] = boto.dynamodb.condition.EQ(value)
+                        dynamo_filter[rkey + '__eq'] = value
 
         limit = 20 if 'limit' not in request.GET else int(request.GET['limit'])
 
+        if esk:
+            dynamo_filter['exclusive_start_key'] = esk
+
         if hash_key_value:
             # do a query, we have hash key filter
-            if self._meta.table.schema.range_key_name and rkey in dynamo_filter:
-                rkc = dynamo_filter[rkey]
-            else:
-                rkc = None
-            _items = self._meta.table.query(hash_key_value,
-                                            range_key_condition=rkc,
-                                            max_results=limit,
-                                            exclusive_start_key=esk)
+
+            selected_indexes = set(request.GET.keys()).intersection(set(self._meta.indexes.values()))
+
+            if selected_indexes:
+                for index_field in selected_indexes:
+                    val = request.GET[index_field]
+                    # TODO there is a bug in boto saying that index field is STRING
+                    # when it is actually a NUMBER. Try to find a solution for this
+                    if type(val) is unicode and val.lower() in ('true', 'false'):
+                        val = 0 if val.lower() == 'false' else 1
+                    dynamo_filter[index_field + '__eq'] = val
+                    for index_name, val in self._meta.indexes.iteritems():
+                        if val == index_field:
+                            break
+                    dynamo_filter['index'] = index_name
+
+            _items = self._meta.table.query(limit=limit,
+                                            **dynamo_filter)
         else:
-            _items = self._meta.table.scan(scan_filter=dynamo_filter,
-                                           max_results=limit,
-                                           exclusive_start_key=esk)
+            _items = self._meta.table.scan(limit=limit,
+                                           **dynamo_filter)
 
         items = [it for it in _items]
 
@@ -265,18 +325,18 @@ class DynamoHashResource(Resource):
             bundle = self.build_bundle(obj=obj, request=request)
             bundles.append(self.full_dehydrate(bundle))
 
-        # generate 'next' URI using last_evaluated_key
-        if not _items.last_evaluated_key:
+        # generate 'next' URI using _last_key_seen
+        if not _items._last_key_seen:
             next_uri = None
         else:
-            next_uri = '/api/%s/%s/?offset_hash=%s' % (kwargs['api_name'], kwargs['resource_name'], _items.last_evaluated_key[0])
+            next_uri = '/api/%s/%s/?offset_hash=%s' % (kwargs['api_name'], kwargs['resource_name'], _items._last_key_seen[self._get_hash().name])
 
             # append hash_key filter to NEXT URL if necessary
             if hkey_in_next:
                 next_uri += '&%s=%s' % (hkey, hash_key_value)
 
-            if self._meta.table.schema.range_key_name:
-                next_uri += '&offset_range=%s' % _items.last_evaluated_key[1]
+            if self._get_range():
+                next_uri += '&offset_range=%s' % _items._last_key_seen[self._get_range().name]
 
             if 'limit' in request.GET:
                 next_uri += '&limit=%s' % request.GET['limit']
@@ -309,7 +369,7 @@ class DynamoRangeDeclarativeMetaclass(DynamoDeclarativeMetaclass):
 
         #ensure range key condition
         if not hasattr(new_class._meta, 'range_key_condition'):
-            setattr(new_class._meta, 'range_key_condition', EQ)
+            setattr(new_class._meta, 'range_key_condition', '__eq')
 
         #ensure a proper delimeter
         if not hasattr(new_class._meta, 'primary_key_delimiter'):
@@ -322,7 +382,7 @@ class DynamoRangeDeclarativeMetaclass(DynamoDeclarativeMetaclass):
         #if the user is asking us to auto-build their primary keys
         if getattr(new_class._meta, 'build_primary_keys', False) == True:
             schema = new_class._meta.table.schema
-            new_class.base_fields[schema.range_key_name] = fields.NumericRangeKeyField(attribute=schema.range_key_name) if schema.range_key_type == 'N' else fields.StringRangeKeyField(attribute=schema.range_key_name)
+            new_class.base_fields[self._get_range().name] = fields.NumericRangeKeyField(attribute=self._get_range().name) if schema.range_key_type == 'N' else fields.StringRangeKeyField(attribute=self._get_range().name)
 
         return new_class
 
@@ -335,14 +395,11 @@ class DynamoHashRangeResource(DynamoHashResource):
 
     def __init__(self, *a, **k):
         super(DynamoHashRangeResource, self).__init__(*a, **k)
-        self._range_key_type = int if self._meta.table.schema.range_key_type == 'N' else str
+        self._range_key_type = int if self._get_range().data_type == 'N' else str
 
 
     def dispatch_detail(self, request, **k):
         """Ensure that the range_key is received in the correct type"""
-
-        if (type(k['range_key']) is str and k['range_key'][-1] == '*') or k['range_key'] == '*':
-            return self.get_list(request, **k)
 
         k['range_key'] = self._range_key_type(k['range_key'])
         return super(DynamoHashRangeResource, self).dispatch_detail(request, **k)
@@ -356,8 +413,8 @@ class DynamoHashRangeResource(DynamoHashResource):
         kwargs = { 'api_name': self._meta.api_name,
                  'resource_name': self._meta.resource_name }
         if bundle:
-            kwargs['hash_key'] = getattr(bundle.obj, self._meta.table.schema.hash_key_name)
-            kwargs['range_key'] = getattr(bundle.obj, self._meta.table.schema.range_key_name)
+            kwargs['hash_key'] = getattr(bundle.obj, self._get_hash().name)
+            kwargs['range_key'] = getattr(bundle.obj, self._get_range().name)
             for key in ('hash_key', 'range_key'):
                 if kwargs[key]:
                     kwargs[key] = str(kwargs[key])
